Rotation fixed (TODO: 90 degrees constraint)

+++ GameEngine/source/graphics/renderer/camera/camera.h
@@ -47,7 +47,7 @@ public:
 
 	inline void Rotate(int _x, int _y)
 	{
-		m_Yaw += static_cast<float>(_x);
+		m_Yaw += _x;
 		m_Pitch += _y;
 		m_IsRotationDirty = true;
 	}
